requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
openai
python-dotenv
